// lib/identity/types.ts
export type FakeProfile = {
  countryCode: string
  countryName: string
  localeLabel: string
  fullName: string
  gender: 'Male' | 'Female'
  dob: string
  age: number
  email: string
  username: string
  password: string
  phone: string
  street: string
  city: string
  state: string
  zip: string
}